# backend/openapi.yaml
openapi: 3.0.0
paths:
...
  /stats/transaction-summary:
    get:
      operationId: StatsController_transactionSummary
      parameters:
        - name: to
          required: false
          in: query
          description: YYYY-MM-DD (inclusive)
          schema: {}
        - name: from
          required: false
          in: query
          description: YYYY-MM-DD (inclusive)
          schema: {}
        - name: type_local
          required: false
          in: query
          schema: {}
        - name: nature_mutation
          required: false
          in: query
          schema: {}
        - name: code_departement
          required: false
          in: query
          schema: {}
        - name: code_commune
          required: false
          in: query
          schema: {}
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                type: object
                properties:
                  transaction_count:
                    type: string
                  min_date_mutation:
                    type: string
                    nullable: true
                    format: date-time
                  max_date_mutation:
                    type: string
                    nullable: true
                    format: date-time
                  median_valeur_fonciere:
                    nullable: true
                  median_price_per_sqm_built:
                    nullable: true
                required:
                  - transaction_count
      tags: &a3
        - stats
  /stats/city-housing-summary:
    get:
      operationId: StatsController_cityHousingSummary
      parameters:
        - name: offset
          required: false
          in: query
          schema:
            type: integer
            default: 0
        - name: limit
          required: false
          in: query
          schema:
            type: integer
            default: 500
        - name: to
          required: false
          in: query
          description: YYYY-MM-DD (inclusive)
          schema: {}
        - name: from
          required: false
          in: query
          description: YYYY-MM-DD (inclusive)
          schema: {}
        - name: code_commune
          required: false
          in: query
          schema: {}
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    code_commune:
                      type: string
                    nom_commune:
                      type: string
                      nullable: true
                    period_month:
                      type: string
                      format: date-time
                    sale_line_count:
                      type: string
                    median_valeur_fonciere:
                      nullable: true
                    median_price_per_sqm_built:
                      nullable: true
                  required:
                    - code_commune
                    - period_month
                    - sale_line_count
      tags: *a3
  /stats/city-equipment-summary:
    get:
      operationId: StatsController_cityEquipmentSummary
      parameters:
        - name: offset
          required: false
          in: query
          schema:
            type: integer
            default: 0
        - name: limit
          required: false
          in: query
          schema:
            type: integer
            default: 500
        - name: code_commune
          required: false
          in: query
          schema: {}
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    code_commune:
                      type: string
                    nom_commune:
                      type: string
                      nullable: true
                    bpe_millesime:
                      type: integer
                      nullable: true
                    equipment_total_count:
                      type: string
                    education_count:
                      type: string
                    health_count:
                      type: string
                    commerce_count:
                      type: string
                    sport_count:
                      type: string
                    other_count:
                      type: string
                  required:
                    - code_commune
                    - equipment_total_count
      tags: *a3
  /stats/city-social-summary:
    get:
      operationId: StatsController_citySocialSummary
      parameters:
        - name: offset
          required: false
          in: query
          schema:
            type: integer
            default: 0
        - name: limit
          required: false
          in: query
          schema:
            type: integer
            default: 500
        - name: code_commune
          required: false
          in: query
          schema: {}
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    code_commune:
                      type: string
                    nom_commune:
                      type: string
                      nullable: true
                    filosofi_millesime:
                      type: integer
                      nullable: true
                    median_income_eur:
                      nullable: true
                    poverty_rate:
                      nullable: true
                    inequality_ratio:
                      nullable: true
                    caf_beneficiary_share:
                      nullable: true
                    housing_aid_share:
                      nullable: true
                    income_d1_eur:
                      nullable: true
                    income_d9_eur:
                      nullable: true
                  required:
                    - code_commune
      tags: *a3
  /stats/iris-social-summary:
    get:
      operationId: StatsController_irisSocialSummary
      parameters:
        - name: offset
          required: false
          in: query
          schema:
            type: integer
            default: 0
        - name: limit
          required: false
          in: query
          schema:
            type: integer
            default: 500
        - name: code_iris
          required: false
          in: query
          schema: {}
        - name: code_commune
          required: false
          in: query
          schema: {}
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    code_iris:
                      type: string
                    code_commune:
                      type: string
                    nom_commune:
                      type: string
                      nullable: true
                    nom_iris:
                      type: string
                      nullable: true
                    filosofi_millesime:
                      type: integer
                      nullable: true
                    median_income_eur:
                      nullable: true
                    poverty_rate:
                      nullable: true
                    inequality_ratio:
                      nullable: true
                    caf_beneficiary_share:
                      nullable: true
                  required:
                    - code_iris
                    - code_commune
      tags: *a3
  /stats/opportunity-score:
    get:
      operationId: StatsController_opportunityScore
      parameters:
        - name: offset
          required: false
          in: query
          schema:
            type: integer
            default: 0
        - name: limit
          required: false
          in: query
          schema:
            type: integer
            default: 500
        - name: to
          required: false
          in: query
          description: YYYY-MM-DD (inclusive)
          schema: {}
        - name: from
          required: false
          in: query
          description: YYYY-MM-DD (inclusive)
          schema: {}
        - name: code_commune
          required: false
          in: query
          schema: {}
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    code_commune:
                      type: string
                    nom_commune:
                      type: string
                      nullable: true
                    period_month:
                      type: string
                      format: date-time
                    sale_line_count:
                      type: string
                    median_valeur_fonciere:
                      nullable: true
                    price_median_per_sqm:
                      nullable: true
                    median_income_eur:
                      nullable: true
                    poverty_rate:
                      nullable: true
                    qpv_share:
                      nullable: true
                    equipment_density:
                      nullable: true
                    transit_accessibility:
                      nullable: true
                    safety_index:
                      nullable: true
                    assumed_interest_rate:
                      nullable: true
                    assumed_term_months:
                      type: integer
                      nullable: true
                    assumed_max_dti:
                      nullable: true
                    borrowing_capacity_eur:
                      nullable: true
                    price_to_capacity_ratio:
                      nullable: true
                    price_score:
                      nullable: true
                    social_mix_score:
                      nullable: true
                    quality_of_life_score:
                      nullable: true
                    composite_score:
                      nullable: true
                  required:
                    - code_commune
                    - period_month
                    - sale_line_count
      tags: *a3
...
info:
  title: Homepedia API
  description: API on top of dbt marts from data_platform.
  version: 1.0.0
  contact: {}
...
servers: []
...